+++ Backend/src/modules/user/user.service.ts
@@ -1,10 +1,11 @@
-import { APIResponse, UpdateAPIResponse, UserDetailsObject } from '@mailsense/types';
+import { ACCOUNT_SYNC_MODE, APIResponse, UpdateAPIResponse, UserDetailsObject, UserSettings } from '@mailsense/types';
 import { Auth0Service } from 'integrations/auth0/auth0.service.js';
 import { Auth0UserDetailsResponse } from 'integrations/auth0/auth0.types.js';
 import { decrypt } from 'shared/utils/index.js';
-import { UserDocument, UserInput } from './user.model.js';
+import { UserDocument, UserInput, UserSettingsDocument } from './user.model.js';
 import { UserRepository } from './user.repository.js';
 import { ChangePasswordSchema, UpdateUserSchema } from './user.schema.js';
+import { UserSettingsRepository } from './user-settings.repository.js';
 
 export class UserService {
     private auth0Service: Auth0Service;
@@ -48,5 +49,32 @@ export class UserService {
         };
         await this.auth0Service.changeUserPassword(auth0UserId, changePasswordBody);
         return { status: true, message: 'Password updated successfully' };
+    }
+
+    public async getUserSettings(auth0UserId: string): Promise<APIResponse<UserSettings>> {
+        let userSettings = await UserSettingsRepository.getUserSettings(auth0UserId);
+        if (!userSettings) {
+            const data = {
+                userId: auth0UserId,
+                account: {
+                    syncSettings: {
+                        globalAutoSync: true,
+                        syncMode: ACCOUNT_SYNC_MODE.CUSTOM_PER_ACCOUNT,
+                        globalSyncInterval: 15,
+                        defaultSyncInterval: 15,
+                    },
+                },
+            };
+            userSettings = await UserSettingsRepository.createUserSettings(data);
+        }
+        return { status: true, message: 'User settings fetched successfully', data: userSettings };
+    }
+
+    public async updateUserSettings(auth0UserId: string, data: UserSettings): Promise<APIResponse<UserSettingsDocument | null>> {
+        const userSettings = await UserSettingsRepository.updateUserSettings(auth0UserId, data);
+        if (!userSettings) {
+            return { status: false, message: 'User settings not found', data: null };
+        }
+        return { status: true, message: 'User settings updated successfully', data: userSettings };
     }
 }